Issue 78: Move LogoutSuccessHandler
